Support string model_selector parameter, move tests
